NOISSUE Move global properties into api module
Global configurations or e.g. names should be accessible to all modules via the api module and not the shared region connector module.
