+++ api/src/main/java/energy/eddie/api/agnostic/GlobalConfig.java
@@ -2,6 +2,7 @@ package energy.eddie.api.agnostic;
 
 public final class GlobalConfig {
     public static final String ERRORS_PROPERTY_NAME = "errors";
+    public static final String ERRORS_JSON_PATH = "$." + ERRORS_PROPERTY_NAME;
 
     private GlobalConfig() {}
 }
