+++ region-connectors/region-connector-at-eda/src/test/java/energy/eddie/regionconnector/at/eda/web/PermissionRequestControllerTest.java
@@ -30,8 +30,8 @@ import java.time.ZoneOffset;
 import java.util.Optional;
 import java.util.stream.Stream;
 
+import static energy.eddie.api.agnostic.GlobalConfig.ERRORS_JSON_PATH;
 import static energy.eddie.regionconnector.shared.web.RestApiPaths.PATH_PERMISSION_STATUS_WITH_PATH_PARAM;
-import static energy.eddie.spring.regionconnector.extensions.RegionConnectorsCommonControllerAdvice.ERRORS_JSON_PATH;
 import static org.hamcrest.Matchers.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
